+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The change log for the API server for Okuna.
 - Add search APIs to administrated, moderated, favorite, joined, subscribed
   communities
 - Add subscribe to user APIs
+- Add trending posts curation and clean up jobs
 - Remove username from forgot password API
 - Fix bug in profile post counts